Added setboundingvolume method
Added setBoundingVolume method to U4DModel. This method, however, needs
to be moved to U4DStaticModel. Also added a Blender transformation
matrix. Need to find a way how to remove this matrix and import all
blender data as OpenGL coordinates. I also have to fix the
setVertexDistanceFromCenterMethod in U4DStaticModel