added new method to tetrahedron
Added a method to get the closest triangle in the tetrahedron to a
point.
Made determineClosestPointOnSImplexToPoint be part of the base class
instead of just the GJK. This way I can use it in other algorithms
In U4DTriangles made two new methods which calculate the distance of
the triangle to a particular point.